% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbeta_plot.R
\name{dbeta_plot}
\alias{dbeta_plot}
\title{Plots probability density function of beta}
\usage{
dbeta_plot(alpha, beta, by = 0.0025)
}
\arguments{
\item{alpha}{Number of prior successes}

\item{beta}{Number of prior failures}

\item{by}{The increment length to approximate the continuous support}
}
\value{
The probability a beta equals x
}
\description{
Plots probability density function of beta
}
\examples{
dbeta_plot(10,5,5)
}
