% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbetbin.R
\name{dbetbin}
\alias{dbetbin}
\title{Probability mass function of a beta binomial}
\usage{
dbetbin(x, n, alpha, beta)
}
\arguments{
\item{x}{An integer less than or equal to n}

\item{n}{The number of trials}

\item{alpha}{Number of prior successes}

\item{beta}{Number of prior failures}
}
\value{
The probability a beta binomial equals x
}
\description{
Calculates probability mass function of beta binomial
}
\examples{
dbetbin(10,20,10,10)
}
