% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbinom_plot.R
\name{dbinom_plot}
\alias{dbinom_plot}
\title{Plots probability mass function of beta binomial}
\usage{
dbinom_plot(n, p)
}
\arguments{
\item{n}{The number of trials}

\item{p}{The probability of success on each trial}
}
\value{
The probability a binomial equals x
}
\description{
Plots probability mass function of beta binomial
}
\examples{
dbinom_plot(10,0.5)
}
