% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbetbin.R
\name{pbetbin}
\alias{pbetbin}
\title{Cumulative distribution function of a beta binomial}
\usage{
pbetbin(x, n, alpha, beta)
}
\arguments{
\item{x}{An integer less than or equal to n}

\item{n}{The number of trials}

\item{alpha}{Number of prior successes}

\item{beta}{Number of prior failures}
}
\value{
The probability a beta binomial is less than or equal to x
}
\description{
Calculates cumulative distribution function of beta binomial
}
\examples{
pbetbin(10,20,10,10)
}
